+++ Molten.Graphics.DX11/Resources/Buffers/TypedBufferDX11.cs
@@ -57,7 +57,7 @@ internal unsafe class TypedBufferDX11 : BufferDX11
             {
                 BufferEx = new BufferexSrv()
                 {
-                    NumElements = ElementCount,
+                    NumElements = (uint)ElementCount,
                     FirstElement = 0,
                     Flags = 0 // See: https://docs.microsoft.com/en-us/windows/win32/api/d3d11/ne-d3d11-d3d11_bufferex_srv_flag
                 },
@@ -77,7 +77,7 @@ internal unsafe class TypedBufferDX11 : BufferDX11
                 ViewDimension = UavDimension.Buffer,
                 Buffer = new BufferUav()
                 {
-                    NumElements = ElementCount,
+                    NumElements = (uint)ElementCount,
                     FirstElement = 0,
                     Flags = 0U, // (uint)BufferUavFlag.None,
                 }
